+++ common/ext/class.GenerisInstaller.php
@@ -17,13 +17,16 @@
  *
  * Copyright (c) 2008-2010 (original work) Deutsche Institut für Internationale Pädagogische Forschung (under the project TAO-TRANSFER);
  *               2009-2012 (update and modification) Public Research Centre Henri Tudor (under the project TAO-SUSTAIN & TAO-DEV);
- *               2013-2014 (update and modification) Open Assessment Technologies SA;
+ *               2013-2020 (update and modification) Open Assessment Technologies SA;
  *
  */
 
-use oat\generis\model\data\ModelManager;
+use oat\generis\model\data\Ontology;
 use oat\generis\model\kernel\persistence\smoothsql\search\ComplexSearchService;
-use oat\oatbox\service\ServiceManager;
+use oat\oatbox\service\exception\InvalidService;
+use oat\oatbox\service\exception\InvalidServiceManagerException;
+use core_kernel_persistence_smoothsql_SmoothModel as SmoothModel;
+use common_cache_Cache as CommonCache;
 
 /**
  * Custom extension installer for generis
@@ -40,8 +43,10 @@ class common_ext_GenerisInstaller extends common_ext_ExtensionInstaller
      * Setup the ontology configuration
      *
      * @access public
-     * @author Joel Bout, <joel.bout@tudor.lu>
-     * @return mixed
+     * @throws common_Exception
+     * @throws common_ext_ExtensionException
+     * @throws common_ext_InstallationException
+     * @throws common_ext_ManifestNotFoundException
      */
     public function install()
     {
@@ -50,6 +55,7 @@ class common_ext_GenerisInstaller extends common_ext_ExtensionInstaller
         }
  
         $this->installLoadDefaultConfig();
+        $this->configureOntology();
         $this->installOntology();
         $this->installRegisterExt();
         
@@ -57,5 +63,26 @@ class common_ext_GenerisInstaller extends common_ext_ExtensionInstaller
         
         $this->log('d', 'Installing custom script for extension ' . $this->extension->getId());
         $this->installCustomScript();
+    }
+
+    /**
+     * @throws common_Exception
+     */
+    protected function configureOntology()
+    {
+        if (!$this->getServiceManager()->has(Ontology::SERVICE_ID)) {
+            $smoothModel = new \core_kernel_persistence_smoothsql_SmoothModel([
+                SmoothModel::OPTION_PERSISTENCE => 'default',
+                SmoothModel::OPTION_READABLE_MODELS =>
+                    [SmoothModel::DEFAULT_WRITABLE_MODEL, SmoothModel::DEFAULT_READABLE_MODEL],
+                SmoothModel::OPTION_WRITEABLE_MODELS =>
+                    [SmoothModel::DEFAULT_WRITABLE_MODEL],
+                SmoothModel::OPTION_NEW_TRIPLE_MODEL =>
+                    SmoothModel::DEFAULT_WRITABLE_MODEL,
+                SmoothModel::OPTION_SEARCH_SERVICE => ComplexSearchService::SERVICE_ID,
+                SmoothModel::OPTION_CACHE_SERVICE => CommonCache::SERVICE_ID
+            ]);
+            $this->getServiceManager()->register(Ontology::SERVICE_ID, $smoothModel);
+        }
     }
 }
