+++ core/kernel/persistence/smoothsql/class.SmoothRdf.php
@@ -25,6 +25,7 @@ use oat\generis\model\OntologyRdfs;
 use oat\oatbox\service\ServiceManager;
 use oat\oatbox\event\EventManager;
 use oat\generis\model\data\event\ResourceCreated;
+use core_kernel_persistence_smoothsql_SmoothModel as SmoothModel;
 
 /**
  * Implementation of the RDF interface for the smooth sql driver
@@ -65,7 +66,7 @@ class core_kernel_persistence_smoothsql_SmoothRdf implements RdfInterface
     public function add(\core_kernel_classes_Triple $triple)
     {
         if (!in_array($triple->modelid, $this->model->getReadableModels())) {
-            $triple->modelid = $this->model->getDefaultReadableModel();
+            $triple->modelid = SmoothModel::DEFAULT_READABLE_MODEL;
         }
         $query = "INSERT INTO statements ( modelId, subject, predicate, object, l_language, epoch, author) VALUES ( ? , ? , ? , ? , ? , ?, ?);";
         $success = $this->getPersistence()->exec($query, [$triple->modelid, $triple->subject, $triple->predicate, $triple->object, is_null($triple->lg) ? '' : $triple->lg, $this->getPersistence()->getPlatForm()->getNowExpression(), is_null($triple->author) ? '' : $triple->author]);
